+++ src/pkg/mumax/engine/torque.go
@@ -30,5 +30,5 @@ type torqueUpdater struct {
 }
 
 func (u *torqueUpdater) Update() {
-	gpu.Torque(u.τ.Array(), u.m.Array(), u.h.Array(), u.α.Array(), u.α.multiplier[0])
+	gpu.Torque(u.τ.Array(), u.m.Array(), u.h.Array(), u.α.Array(), float32(u.α.multiplier[0]))
 }
